stash test

+++ src/main/java/com/Spring/SpringWebDemo/controller/StudentController.java
@@ -1,13 +1,20 @@
 package com.Spring.SpringWebDemo.controller;
 
-import com.Spring.SpringWebDemo.service.StudentService;
-import com.Spring.SpringWebDemo.model.Student;
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
 
-import java.util.List;
+import com.Spring.SpringWebDemo.model.Student;
+import com.Spring.SpringWebDemo.service.StudentService;
 
 @RestController
 public class StudentController {
@@ -17,9 +24,9 @@ public class StudentController {
 
     @DeleteMapping("/student/{id}")
     public ResponseEntity<String> deleteStudent(@PathVariable int id){
-        /*if(studentService.getStudentById(id) == null)
+        if(studentService.getStudentById(id) == null)
             return new ResponseEntity<>("Student not found",HttpStatus.BAD_REQUEST);
-        else*/
+        else
             return new ResponseEntity<>(studentService.deleteStudent(id), HttpStatus.OK);
